// src/types/Escala.ts
export interface Escala {
  id: number;
  diaSemana: string;
  horaInicio: string;
  horaFim: string;
  intervaloMinutos: number;
}

export interface EscalaFormData {
  diaSemana: string;
  horaInicio: string;
  horaFim: string;
  intervaloMinutos: number;
}

export type DiaSemana = 
  | 'SEGUNDA'
  | 'TERCA'
  | 'QUARTA'
  | 'QUINTA'
  | 'SEXTA'
  | 'SABADO'
  | 'DOMINGO';

export interface HorarioDisponivel {
  hora: string;
  disponivel: boolean;
}